refactor to move some utils functions and fixed imports accordingly
