+++ .gitignore
@@ -57,7 +57,7 @@ OldWorks
 MANIFEST
 logs
 zprivate
-scratch*
+localsandpit
 localdir
 tmp
 temp
