+++ mtpy/gui/SmartMT/visualization/mt_response.py
@@ -11,10 +11,10 @@
 
 from mtpy.gui.SmartMT.Components.PlotParameter import Ellipse, Arrow, StationSelection, Rotation
 from mtpy.gui.SmartMT.gui.plot_control_guis import PlotControlMTResponse
-from mtpy.gui.SmartMT.utils.matplotlib_utils import get_next_fig_num
 from mtpy.gui.SmartMT.visualization.visualization_base import VisualizationBase
 from mtpy.imaging.plot_mt_response import PlotMTResponse
 from mtpy.imaging.plotnresponses import PlotMultipleResponses
+from mtpy.utils.matplotlib_utils import get_next_fig_num
 
 
 class MTResponse(VisualizationBase):
