+++ mtpy/gui/SmartMT/visualization/strike.py
@@ -11,9 +11,9 @@
 from mtpy.gui.SmartMT.Components.FigureSetting import Font
 from mtpy.gui.SmartMT.Components.PlotParameter import FrequencyTolerance, Rotation
 from mtpy.gui.SmartMT.gui.plot_control_guis import PlotControlStrike
-from mtpy.gui.SmartMT.utils.matplotlib_utils import get_next_fig_num
 from mtpy.gui.SmartMT.visualization import VisualizationBase
 from mtpy.imaging.plotstrike import PlotStrike
+from mtpy.utils.matplotlib_utils import get_next_fig_num
 
 
 class Strike(VisualizationBase):
