+++ tests/modeling/ModEM/test_data.py
@@ -6,7 +6,7 @@ from unittest import TestCase
 
 import matplotlib.pyplot as plt
 
-from examples.create_modem_input import select_periods
+from mtpy.core.edi_collection import EdiCollection
 from mtpy.modeling.modem import Data
 # patch that changes the matplotlib behaviour
 from tests import make_temp_dir, plt_wait, plt_close
@@ -90,7 +90,7 @@ def _test_gen(edi_path, error_type_tipper, error_type_z, comp_error_type):
             self.skipTest("edi path does not exist: {}".format(edi_path))
 
         edi_list = glob.glob(edi_path + '/*.edi')
-        period_list = select_periods(edi_list)
+        period_list = EdiCollection(edi_list).select_periods()
         datob = Data(edi_list=edi_list,
                      inv_mode='1',
                      period_list=period_list,
